Checking in basic FFmpeg import library generator project and public header files.
Similar to V8Bindings_prebuild, ffmpeg.vcproj is a Utility project that calls a .bat file to build stuff.  Whenever we need to link to an FFmpeg function, we simply add it to the corresponding .def file and we're set!

Review URL: http://codereview.chromium.org/23021

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ea5b8bb0dd358836695f6e8521be995adbe6e54c