Update the Pepper APIs to the latest spec for the 2D demo plugin.
This also adds the npapi headers to the npapi.gyp file since I got tired of
Visual Studio not finding the files.

This removes the "open file in sandbox" feature which it doesn't look like we
will use.

One more significant change is that I changed to including pepper.h in all
cases, even when pepper is disabled. We used to have a forward declare in
npapi.h for the structs in question, but we'll be adding a lot more structs for
the different contexts and I don't think this will scale. I think its OK fo
rthe pepper API declarations to be available when Pepper isn't enabled.

BUT=none
TEST=none

Review URL: http://codereview.chromium.org/453015

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 66b4e8dfc5cce75c998889d5b968ecb72a204447