Convert to SQLite amalgamation.
SQLite generally distributes using a script-generated "Giant file of
stuff", called the amalgamation.  It's apparently faster, due to more
opportunities for optimization, but for Chromium the attraction is
that we don't need as many patches to SQLite.  Imports of future
versions of SQLite should be simpler, so long as an Ubuntu box is
available.

BUG=71731
TEST=automated testing and watching closely.

Review URL: http://codereview.chromium.org/6835006

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: fdd072ff1b52f405ed3c2b1cfb86c6e92e5018c1