Exclude ffmpeg from the build when building for Android.
Conditionalize the ffmpeg dependency and exclude files which rely on it
for the media and media_unittest targets.  Disable the ffmpeg_unittests,
ffmpeg_tests and media_bench for Android targets as they require it.

This change also conditionalizes the inclusion of yasm_compile.gyp, as
yasm is only used for the ia32 and x86 target architectures.

BUG=
TEST=

Review URL: http://codereview.chromium.org/7977023

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 23f991a603b837fbff921270a2998872f1eefb1e