Add support for GL_COMMANDS_ISSUED_CHROMIUM fence like query.
I'm not sure this is the right way to do this. I started by adding
a few switch/case statements and then realized that there is
no guarntee that glGenQueriesXXX exists which means I have to NOT
call that for GL_COMMANDS_ISSUED_CHROMIUM queries. I ended up
going down the path of a base class and this is where it lead.

TEST=unit tests
BUG=117768


Review URL: http://codereview.chromium.org/9694025

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c45f197ed2d12b4f16a3d3ab05040d7c58e479b0