Addition of a set of basic unit tests for the GL_CHROMIUM_copy_texture extension.
This change also introduces a new parameter to the extension, internal_format.  This is to bring the extension closer in functionality to glCopyTexImage.

The tests uncovered state that was not being properly restored:  GL_BLEND.  This is also corrected.

TEST=GLCopyTextureCHROMIUMTest.*
BUG=126178

Review URL: http://codereview.chromium.org/10356004

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0a1e9ad53f2f7f2fc6139644da122e272b88fb80