Roll libphonenumber r456
This also changes libphonenumber.gyp. A separate library with no metadata needs
to be built. This allows us to build a production library with production
metadata and the unit tests with testing metadata. Before that it was unclear
which metadata was used for testing.

TEST=libphonenumber_unittests


Review URL: https://chromiumcodereview.appspot.com/10412031

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a3e40033f711e3b8414b491827cfb6c35b9b8e53