Fix misaligned address in sqlite3.c
Fix the problem that an int* pointer is assigned with a value calculated
from a UChar* pointer and an offset. If the offset is odd, the int* pointer
will not be 4-byte aligned which causes SIGBUS on some CPUs.

Please see #9 and #10 in crbug.com/151673 for details.

BUG=151673


Review URL: https://chromiumcodereview.appspot.com/11183042

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d567d600546768dbe7bff205ef8e84d9c660c9dd