Specify dependences correctly when building mac tools for iOS.
Several auxiliary tools used when building Chrome for iOS are generated via the
following two-step pattern: (1) compiling the executable with ninja, and (2)
copying the executable into a location that is shared with other projects.
Previously, these two steps were specified as actions in the same target.
However, as the ordering of multiple actions in one target is defined only by
inputs and outputs and the compilation action had no inputs/outputs set, the
copying action wasn't properly depending on the compilation action.  

As it's challenging to set correct inputs for the ninja build, this CL
separates these actions into two targets in order to be able to specify that
the second action should not run until the first action finishes.


Review URL: https://chromiumcodereview.appspot.com/11301003

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5b6413f7eaf2c41cded749e442d12bb9aa24bf84