add a dependency on -lz to freetype2.
It turns out that the lucid system freetype looks like it was compiled
with a dependency on the system zlib (unsurprisingly). In the custom build
I just added, I missed declaring the link library dependency on -lz.

I'm not sure if this will correctly interact w/ use_system_zlib, but this
appears to reproduce the existing correct build properly (which depends
on the system zlib even if use_system_zlib is set to zero), so perhaps
use_system_zlib is broken for DRT?

R=tony@chromium.org
BUG=


Review URL: https://chromiumcodereview.appspot.com/12316085

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1b2298d643e195aff25ced72f4ce4c38f475377b