Adds TCMalloc support for Android.
This is part of the effort to bring TCMalloc to android.
The first goal is to get instrumentation to facilitate 
integration with DMP and memory profiling.
This is not yet intended for full production usage as the
default allocator.

BUG=162208

Review URL: https://chromiumcodereview.appspot.com/14321006

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3c8fe548ff6db60540aa8ee976801b10aa5a3d0f