Removed patch to GCC-style assembly in third_party/lzma_sdk.
Previously, this code caused a segmentation fault under certain compilers using optimization; in particular the high 32 bits of EBX are zeroed as a result of the mov instruction. However, if the compiler is using EBX to store a 64-bit address (which proved to be the case for Courgette), this results in a segmentation fault.

The following fixes were also considered and verified as working:
`push $ebx` at the start of the modified assembly section and `pop $ebx` at the end.
Add "ebx" to the clobber list of the instruction to prevent the compiler from depending upon it not being clobbered.

However, I can't seem to verify that the build breaks on Linux x86 without this change; so I don't see why we need this patch. Therefore, I've simply removed it.

BUG=248385
TBR=bashi@chromium.org

Review URL: https://chromiumcodereview.appspot.com/16402015

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b5f5d16b76dd83cb9bc7da539f9574e0ed4025cb