Avoid register corruption in LZMA ASM in 64-bit architectures.
I tried this previously but missed that $ebx is used to track the global offset table in position-independent-code.

I wish the .patch file was smaller: the actual change is much smaller, but I can't seem to make `diff` spit out a better patch at me, and I don't want to hand-edit it out of fear of breaking the patching.

More on the reasons why this patch is necessary are at the bug.

BUG=248385

Review URL: https://chromiumcodereview.appspot.com/16965016

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f6820ba92b1bd5f3f44cbedb116e8593dd6e5968