Use the system stdint.h on Windows in leveldatabase.
Skia wishes to use the system stdint.h on Windows. This is now possible
since VC2010 and later provide a stdint.h. VC2008 has been deprecated for
building Chromium (and Skia) for a full year now.

There are a number of stdint.h implementations in the codebase, some of
which hide others, and most of which are not complete (Skia needs
uintmax_t which most do not provide).

The stdint.h in leveldatabase/src/port/win is incomplete and is also
providing (apparently by accident) the stdint.h for Chromium's
src/content. As such, it is also hiding the system stdint.h from Skia
header files included in src/content.

The stdint.h in leveldatabase/src/port/win is also the only file in that
directory, so this change simply removes this directory from the list of
includes.

BUG=81490
R=jsbell@chromium.org

Review URL: https://codereview.chromium.org/27089002

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7b9e7df1604df77026dd9184927ed64b27f2ec6c