Zip utf-8 file names with encoding bit set.
This commit sets the 'locale' bit in global directory of a zip which
means that the file name was written in UTF-8. Chromium always writes
and reads zips as UTF-8 so it is not a problem per se but other, more
zip compliant, programs see packed files with wrong paths if this bit
is not set.

BUG=325737

Review URL: https://codereview.chromium.org/94093003

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b9b88b9aa86bdaa3c9f053877a4c1e9562652f04