[rac] Add AddressValidator API.
This patch adds the AddressValidator API without the implementation. It is
intended to start integration into Chrome in parallel with implementing
the API in the library. This is not the final API, as it may change as
integration into Chrome progresses.

BUG=327046

Review URL: https://codereview.chromium.org/106863006

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 95598bcb34a10b1c7fdd2d3e6454f84c72509841