Allow braille translation result to be larger than twice the size of the input.
The liblouis wrapper used to allocate an output buffer that was twice as big
as the number of input characters.  There are situations where this is
not enough.  The new approach is to retry with a larger buffer up to
a certain limit.

R=dtseng@chromium.org
BUG=none

Review URL: https://codereview.chromium.org/124403002

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a0765c2c4cbf30ccf0ed282369d2d104756d1956