Make it possible to use OSMesa on Android
This patch will start compiling OSMesa as a dependency of the
content_shell_apk target, and changes the GL implementation for Android
to recognize OSMesa as a valid implementation.

libosmesa.so will be separately included in ContentShell.apk, and
will be lazily loaded depending on whether OSMesa will be used or not.

When running a set of 54 layout tests on a Nexus 4, total time is
[69, 68, 68] ~68 seconds with this patch, [62, 60, 61] ~61 seconds
without this patch. Compensated for the setup time (13 seconds),
this means pixel tests are on average 14 percent slower. I believe
the win of (a) consistency with Chrome, (b) more reliable results and
(c) less cross-device differences warrants this, however.

Switching to OSMesa does also enable us to reliably enable impl-side
painting without getting garbage rendering on higher-end devices. At
time time it's not yet enabled by default however, as it makes
content_shell time out when ran on the Nexus 7.

BUG=232044, 248925, 250777

Review URL: https://codereview.chromium.org/23868030

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: abeda1fe53a5f2763cc17d45d53316b5ee27b510