Update clang-format binaries and scripts to clang revision 202065 (24 February 2014).
I followed these instructions:
https://code.google.com/p/chromium/wiki/UpdatingClangFormatBinaries

STATISTICS from an analysis of the 1,957 most recent clang-formattable
commits to Chrome (~20 days of activity):

== How well did the old
== version of clang-format do?

43.33% commits were completely clang-format "clean" according to the
old version. This metric is an indicator of either clang-format
matching the prevailing style (which is more likely to happen with
smaller CLs), or people using clang-format as part of their workflow.

Between 77.30% and 87.18% of all committed lines (depending on how you
count -- the lower bound is very conservative) were clang-format
"clean" according to the old version. This metric is an indicator of
the deviations between prevailing style and the auto-formatting of
clang-format.

== How different is the new version of clang-format from the old
== version?

0.36% of lines in all commits would have been formatted differently
between the new and old versions of the tool.

10.94% commits (regardless of whether they were clean) would have had
at least one line formatted differently by the new version of clang-
format compared to the old version.

== What do the differences look like, qualitatively?

I've uploaded a CL that shows the formatting differences here:
https://codereview.chromium.org/179853003/ . The most common
differences I see are: [a] increased indentation when a unary operator
like ! precedes the a wrapped function call (so that parameters are +4
relative to the start of the function name) [b] More likely to wrap
"const OVERRIDE" on a function signature and [c] Improved alignment of
parameters that are inside #ifdefs. All of these are expected given
the description in the upstream changelog.

BUG=344916
R=ajwong@chromium.org

Review URL: https://codereview.chromium.org/180023007

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4f5120b7485af80deec0c78953873dc5c006ead7